+++ changes.md
@@ -1,2 +1,3 @@
-- Pre-import after selecting skin to reduce inject time
-- New inject system (no need to download dll anymore, faster inject, no more window bug)
+- Added Enable/Disable button for bocchi mini window
+- Disabled dodge button since riot patched it
+- QoL changes